// apps/iroha-web/src/lib/postgrest/types.ts
export interface paths {
// ...
    "/rpc/public_summary": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        put?: never;
        post: {
            parameters: {
                query?: never;
                header?: {
                    /** @description Preference */
                    Prefer?: components["parameters"]["preferParams"];
                };
                path?: never;
                cookie?: never;
            };
            requestBody: {
                content: {
                    "application/json": Record<string, never>;
                    "application/vnd.pgrst.object+json;nulls=stripped": Record<string, never>;
                    "application/vnd.pgrst.object+json": Record<string, never>;
                    "text/csv": Record<string, never>;
                };
            };
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        activities: {
            /**
             * Format: uuid
             * @description Note:
             *     This is a Primary Key.<pk/>
             */
            id?: string;
            /** Format: text */
            sport_type?: string;
            /** Format: text */
            title?: string;
            /** Format: timestamp with time zone */
            started_at?: string;
            /** Format: timestamp with time zone */
            ended_at?: string;
            /** Format: text */
            timezone?: string;
            /** Format: numeric */
            distance_m?: number;
            /** Format: int32 */
            duration_s?: number;
            /** Format: int32 */
            moving_time_s?: number;
            /** Format: numeric */
            elevation_gain_m?: number;
            /** Format: int32 */
            avg_hr?: number;
            /** Format: int32 */
            max_hr?: number;
            /** Format: numeric */
            avg_pace_s_per_km?: number;
        };
        public_activities: {
            /** Format: text */
            id?: string;
            /** Format: text */
            sport_type?: string;
            /** Format: text */
            title?: string;
            /** Format: timestamp with time zone */
            started_at?: string;
            /** Format: timestamp with time zone */
            ended_at?: string;
            /** Format: text */
            timezone?: string;
            /** Format: numeric */
            distance_m?: number;
            /** Format: int32 */
            duration_s?: number;
            /** Format: int32 */
            moving_time_s?: number;
            /** Format: numeric */
            elevation_gain_m?: number;
            /** Format: int32 */
            avg_hr?: number;
            /** Format: int32 */
            max_hr?: number;
            /** Format: numeric */
            avg_pace_s_per_km?: number;
        };
    };
    responses: never;
    parameters: {
        /** @description Preference */
        preferParams: string;
        /** @description Preference */
        preferReturn: "return=representation" | "return=minimal" | "return=none";
        /** @description Preference */
        preferCount: "count=none";
        /** @description Preference */
        preferPost: "return=representation" | "return=minimal" | "return=none" | "resolution=ignore-duplicates" | "resolution=merge-duplicates";
        /** @description Filtering Columns */
        select: string;
        /** @description On Conflict */
        on_conflict: string;
        /** @description Ordering */
        order: string;
        /** @description Limiting and Pagination */
        range: string;
        /** @description Limiting and Pagination */
        rangeUnit: string;
        /** @description Limiting and Pagination */
        offset: string;
        /** @description Limiting and Pagination */
        limit: string;
        "rowFilter.activities.id": string;
        "rowFilter.activities.sport_type": string;
        "rowFilter.activities.title": string;
        "rowFilter.activities.started_at": string;
        "rowFilter.activities.ended_at": string;
        "rowFilter.activities.timezone": string;
        "rowFilter.activities.distance_m": string;
        "rowFilter.activities.duration_s": string;
        "rowFilter.activities.moving_time_s": string;
        "rowFilter.activities.elevation_gain_m": string;
        "rowFilter.activities.avg_hr": string;
        "rowFilter.activities.max_hr": string;
        "rowFilter.activities.avg_pace_s_per_km": string;
        "rowFilter.public_activities.id": string;
        "rowFilter.public_activities.sport_type": string;
        "rowFilter.public_activities.title": string;
        "rowFilter.public_activities.started_at": string;
        "rowFilter.public_activities.ended_at": string;
        "rowFilter.public_activities.timezone": string;
        "rowFilter.public_activities.distance_m": string;
        "rowFilter.public_activities.duration_s": string;
        "rowFilter.public_activities.moving_time_s": string;
        "rowFilter.public_activities.elevation_gain_m": string;
        "rowFilter.public_activities.avg_hr": string;
        "rowFilter.public_activities.max_hr": string;
        "rowFilter.public_activities.avg_pace_s_per_km": string;
    };
// ...
}